Add files via upload
This script scrapes DDW PFAS data from the Water Boards website and determines which water systems have PFAS detections (could easily be changed to be detections above the notification level, or some other level). Then, the bio-monitoring data from participants in water systems with detections and without detections are compared in a Wilcox rank sum test. For this case, I randomly assigned each participant to a water system, so no associations are identified.